refactor: common한 RollingpaperListItem 작성
